Lab3/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
redis==5.0.1
bcrypt==4.1.2

